Hide unpreviliged links
In iPlug SE a new role "instanceAdmin" is introduced (see
https://redmine.informationgrid.eu/issues/979). Users that have the role
"instanceAdmin" are not allowed to do anything but change some aspects
of the instance they have the right for.
To keep the known behavior all links are displayed for anonymous users
and the "admin" (role "admin"). For users with other roles, most links
are hidden.